+++ src/css-rules/no-empty-blocks/node-remover.ts
@@ -8,11 +8,11 @@ import type { TSESTree } from '@typescript-eslint/utils';
  * @param fixer The ESLint fixer.
  * @returns The fix object.
  */
-export function removeNodeWithComma(ruleContext: Rule.RuleContext, node: TSESTree.Node, fixer: Rule.RuleFixer) {
+export const removeNodeWithComma = (ruleContext: Rule.RuleContext, node: TSESTree.Node, fixer: Rule.RuleFixer) => {
   const sourceCode = ruleContext.sourceCode;
   const tokenAfter = sourceCode.getTokenAfter(node as Rule.Node);
   if (tokenAfter && tokenAfter.value === ',' && node.range && tokenAfter.range) {
     return fixer.removeRange([node.range[0], tokenAfter.range[1]]);
   }
   return fixer.remove(node as Rule.Node);
-}
+};